move createBuffer next to Tile move also the test
