+++ include/dlaf/tile.h
@@ -15,6 +15,7 @@
 #include "dlaf/memory/memory_view.h"
 #include "dlaf/types.h"
 #include "dlaf/util_math.h"
+#include "dlaf/common/buffer.h"
 
 namespace dlaf {
 
@@ -170,6 +171,12 @@ private:
   using Tile<const T, device>::ld_;
   using Tile<const T, device>::p_;
 };
+
+/// @brief Create a common::Buffer from a Tile
+template <class T, Device device>
+auto create_buffer(const dlaf::Tile<T, device>& tile) {
+  return dlaf::common::Buffer<T*>(tile.ptr({0, 0}), tile.size().cols(), tile.size().rows(), tile.ld());
+}
 
 #include <dlaf/tile.ipp>
 }
