+++ test/unit/CMakeLists.txt
@@ -15,7 +15,6 @@ DLAF_addTest(test_matrix SOURCES test_matrix.cpp USE_MAIN HPX)
 DLAF_addTest(test_blas_tile SOURCES test_blas_tile.cpp USE_MAIN HPX)
 DLAF_addTest(test_lapack_tile SOURCES test_lapack_tile.cpp USE_MAIN HPX)
 DLAF_addTest(test_tile SOURCES test_tile.cpp USE_MAIN HPX)
-DLAF_addTest(test_util_tile SOURCES test_util_tile.cpp USE_MAIN HPX)
 
 add_subdirectory(common)
 add_subdirectory(communication)
